+++ src/components/image_components/ImageFlexCenter.tsx
@@ -8,6 +8,7 @@ interface Props {
   imageCategory: string;
   width: number;
   height?: number;
+  marginTop?: number;
 }
 
 export const ImageFlexCenter: React.FC<Props> = ({
@@ -17,9 +18,10 @@ export const ImageFlexCenter: React.FC<Props> = ({
   imageCategory,
   width,
   height,
+  marginTop,
 }) => {
   return (
-    <Flex justifyContent="center" alignItems="center" mt={3}>
+    <Flex justifyContent="center" alignItems="center" mt={marginTop || 3}>
       <Image
         src={`https://res.cloudinary.com/dg2r37ygd/image/upload/${imageSize}/v1605448126/images/${imageCategory}/${src}.jpg`}
         alt={altText}
