+++ src/containers/InterviewsPage.tsx
@@ -1,16 +1,16 @@
 import React, { useState, useEffect } from 'react';
-import { Flex, Text, Box, Link, Tag, Image } from '@chakra-ui/react';
+import { Flex, Text, Box, Link, Heading } from '@chakra-ui/react';
 import NextLink from 'next/link';
-import { FrontMatterArticlesType } from '../types/types';
-import EmailSubscription from './partials/EmailSubscription';
-import ArticlePreviewCard from './cards/ArticlePreviewCard';
+import { FrontMatterInterviewsType } from '../types/types';
+import EmailSubscription from '../components/partials/EmailSubscription';
+import InterviewPreviewCard from '../components/cards/InterviewPreviewCard';
 import { sortNumberByPublishedDateRemote } from '../helpers/sortNumberByPublishedDate';
 
-export default function ArticlesPage({ articlePosts }) {
+export default function InterviewsPage({ interviewPosts }) {
   const [sortedPosts, setSortedPosts] = useState([]);
 
   const handleSortingOrder = () => {
-    return articlePosts.sort(sortNumberByPublishedDateRemote);
+    return interviewPosts.sort(sortNumberByPublishedDateRemote);
   };
 
   useEffect(() => {
@@ -19,7 +19,10 @@ export default function ArticlesPage({ articlePosts }) {
 
   return (
     <Flex mt={12} w="100%" direction="column" alignItems="center">
-      <Box as="section" mt={12} px={[4, 8]}>
+      <Box as="section" px={[4, 8]}>
+        <Heading as="h1" size="xl" mb={4} color="cyan.900">
+          Interviews
+        </Heading>
         <Flex
           pb={[4, 8]}
           w="100%"
@@ -35,21 +38,21 @@ export default function ArticlesPage({ articlePosts }) {
                   <Box flex={['1 0 100%']} mb={8} key={frontMatter.data.slug}>
                     <NextLink
                       passHref
-                      href={`/articles/${frontMatter.data.slug}`}
+                      href={`/interviews/${frontMatter.data.slug}`}
                     >
                       <Link>
                         <Flex
                           justifyContent={['center', 'space-between']}
                           alignItems="center"
                         ></Flex>
-                        <ArticlePreviewCard
+                        <InterviewPreviewCard
                           slug={frontMatter.data.slug}
+                          title={frontMatter.data.title}
                           imageUniqueIdentifier={
                             frontMatter.data.imageUniqueIdentifier
                           }
-                          title={frontMatter.data.title}
                           description={frontMatter.data.description}
-                          readingTime={frontMatter.data.readingTime}
+                          // readingTime={frontMatter.data.readingTime}
                         />
                       </Link>
                     </NextLink>
@@ -58,11 +61,12 @@ export default function ArticlesPage({ articlePosts }) {
             )}
           {sortedPosts.length < 1 && (
             <Text fontSize="4xl" color="cyan.500">
-              No Articles Found. Coming soon...
+              No Interviews Found. Coming soon...
             </Text>
           )}
         </Flex>
       </Box>
+
       <Box as="section" px={[4, 8]}>
         <EmailSubscription />
       </Box>
