+++ src/pages/articles/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import Head from 'next/head';
 import PageContainer from '../../containers/PageContainer';
-import ArticlesPage from '../../components/ArticlesPage';
+import ArticlesPage from '../../containers/ArticlesPage';
 
 import { articleFilePaths, ARTICLE_PATH } from '../../utils/mdxUtils';
 import fs from 'fs';
